requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
six==1.10.0
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
click==6.7
Flask==0.12.1
imutils==0.4.2
itsdangerous==0.24
Jinja2==2.9.6
jsonpickle==0.9.4
keyboard==0.9.13
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
opencv-contrib-python==3.2.0.7
opencv-python==3.2.0.7
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
virtualenv==15.1.0
Werkzeug==0.12.1
whitenoise==3.3.0
xmljson==0.1.8
djangorestframework==3.6.3
requests==2.9.1
